server/requirements.txt
fastapi[standard]==0.113.0
pydantic==2.8.0
sqlmodel
pandas
psycopg2-binary
gensim
openpyxl
python-dotenv